hq_backend/api/requirements_api.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-cloud-storage==2.14.0
supabase==2.3.4
openai==1.12.0
python-dotenv==1.0.1
langchain==0.1.6
PyPDF2==3.0.1
